requirements.txt
cloudscraper>=1.2.71
psycopg2-binary>=2.9.9
